Avoid occasional failure to get “next” track playing when playing a new album or track
cmus queues tracks in a background process, and so the ”next” track (our new first track) may not be ready to move to when the ‘next’ command is sent. We work around this for now by waiting for 100ms.

This also switches to using a fresh playlist instead of the play queue.